Refactor application code and update logger dependencies
Removed old implementations of Cipher and Runner classes and refactored them into new, more efficient designs. This made the application code easier to navigate and manage. In addition to this, a missing 'log4j-core' dependency was added to pom.xml to ensure proper logging during runtime. The interface `CipherStrategy` was also updated for better integration with new Runner code.